fix: sync namespace
