+++ Ticket.Application/Queries/TicketPlus/GetAccessToken/GetAccessTokenDto.cs
@@ -1,4 +1,4 @@
-﻿namespace Ticket.Application.Queries.TicketPlus.GetAccessTokenQuery;
+﻿namespace Ticket.Application.Queries.TicketPlus.GetAccessToken;
 
 public class GetAccessTokenDto
 {
