+++ Ticket.Application/Queries/TicketPlus/GetCaptchaAnswer/GetCaptchaAnswerDto.cs
@@ -1,4 +1,4 @@
-﻿namespace Ticket.Application.Queries.TicketPlus.GetCaptchaAnswerQuery;
+﻿namespace Ticket.Application.Queries.TicketPlus.GetCaptchaAnswer;
 
 public class GetCaptchaAnswerDto
 {
