+++ Ticket.Application/Queries/TicketPlus/GetS3ProductInfo/GetS3ProductInfoDto.cs
@@ -1,6 +1,6 @@
 ﻿using Ticket.Domain.Entities.TicketPlus;
 
-namespace Ticket.Application.Queries.TicketPlus.GetProductInfoQuery;
+namespace Ticket.Application.Queries.TicketPlus.GetS3ProductInfo;
 
 public class GetS3ProductInfoDto
 {
